ThesisScraper/requirements.txt
patchright
tiktoken
anthropic
openai
google-genai
